@@ -2,39 +2,36 @@
   "name": "digitalsathi",
   "version": "0.1.0",
   "private": true,
+ 
   "scripts": {
-    "dev": "next dev",
+    "dev": " next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "debug": "ndb server.js"
-  },
+    "lint": "next lint"
+   },
   "dependencies": {
-    "@hookform/resolvers": "^3.3.2",
     "@tinymce/tinymce-react": "^4.3.2",
     "autoprefixer": "10.4.16",
-    "axios": "^1.6.2",
     "bcryptjs": "^2.4.3",
+    "cookie": "^0.6.0",
+    "dotenv": "^16.3.2",
     "eslint": "8.55.0",
-    "eslint-config-next": "14.0.4",
+    "eslint-config-next": "^14.1.0",
     "express": "^4.18.2",
+    "js-cookie": "^3.0.5",
     "jsonwebtoken": "^9.0.2",
     "mongodb": "^6.3.0",
     "mongoose": "^8.0.3",
-    "mongoose-unique-validator": "^4.0.0",
     "multer": "^1.4.5-lts.1",
     "ndb": "^1.1.5",
-    "next": "14.0.4",
+    "next": "^14.1.0",
     "nodemon": "^3.0.2",
     "postcss": "8.4.32",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-hook-form": "^7.49.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.12.0",
     "react-toastify": "^9.1.3",
     "tailwindcss": "3.3.6",
-    "uuid": "^9.0.1",
-    "validator": "^13.11.0",
-    "yup": "^1.3.3"
+    "validator": "^13.11.0"
   }
 }
