@@ -1,5 +1,6 @@
 import mongoose from "mongoose";
-
+ 
+ 
 const connect = async () => {
     try{
             await mongoose.connect(process.env.MONGO_URL, {
@@ -12,3 +13,5 @@ const connect = async () => {
     }
 }
 export default connect;
+
+ 
