@@ -1,23 +1,30 @@
 import { NextResponse } from "next/server";
 import connect from "../../../../server";
 import Categories from "../../../../models/Categories";
+import Users from "../../../../models/Users";
 
-export const GET = async (request) => {
+export const GET = async (request, response, next) => {
+  
   try{
 
+      const url = new URL(request.url);
+      const userId=url.searchParams.get('userId');
       await connect ();
-      const posts = await  Categories.find()
+      let user = await Users.findOne({_id:userId});
+      let posts = await  Categories.find();
+      posts = posts.filter(a=> user.usrRole == 'ADMIN' || a.userId == userId);
       return new NextResponse (JSON.stringify(posts), {status: 200});
 
   }catch(error){
-    return new NextResponse ("Erron while fetching data" + error, {status: 500});
+    return new NextResponse ("Erron while fetching data: " + error, {status: 500});
   }
-}
+};
 
 export  const  POST = async (request) =>{ 
   try  
     {
-      const {catName} = await request.json(); 
+ 
+      const {catName, userId} = await request.json(); 
       await connect ();
 
       const existingCatName = await Categories.findOne({catName});
@@ -27,7 +34,7 @@ export  const  POST = async (request) =>{
       }
       else
       {
-        const category = new Categories({catName});
+        const category = new Categories({catName, userId});
         const result = await category.save();
         return NextResponse.json({result, success:true}, {status: 200});
       }
@@ -35,9 +42,7 @@ export  const  POST = async (request) =>{
       if (error.name === 'ValidationError') {
         const messages = Object.values(error.errors).map(val => val.message);
         return NextResponse.json({ success: false, message: messages }, {status:400});
-      }
-      else
-      {
+      }else{
         return new NextResponse ("Erron while saving data" + error, {status: 400});
       }
     }
