@@ -4,18 +4,21 @@ import { Courses } from "../../../../models/Courses";
 import Users from "../../../../models/Users";
 import { PROTECT } from "../login/route";
 
-export const GET =  async (request, response, next) => {
+
+export const GET = async (request, response, next) => {
   
   try{
-     
+
+      const url = new URL(request.url);
+      const userId=url.searchParams.get('userId');
       await connect ();
-      // let user=await Users.findOne({_id:'65a62bdcba316ba0dc059e30'});
+      let user=await Users.findOne({_id:userId});
       let posts = await  Courses.find();
-      // posts=posts.filter(a=> user.usrRole=='ADMIN' || a.userId=='65a62bdcba316ba0dc059e30');
+      posts=posts.filter(a=> user.usrRole == 'ADMIN' || a.userId == userId);
       return new NextResponse (JSON.stringify(posts), {status: 200});
 
   }catch(error){
-    return new NextResponse ("Erron while fetching data" + error, {status: 500});
+    return new NextResponse ("Erron while fetching data: " + error, {status: 500});
   }
 };
 
@@ -23,10 +26,10 @@ export  const  POST = async (request) =>{
   
     try    
     {
-      const {coName, coTags, coIntro, coDesc, coCat, coVal, coPrice, coDisc, coImage } = await request.json(); 
+      const {coName, coTags, coIntro, coDesc, coCat, coVal, coPrice, coDisc, coImage, userId } = await request.json(); 
       await connect ();
 
-      const posts = new Courses({coName, coTags, coIntro, coDesc, coCat, coVal, coPrice, coDisc, coImage });
+      const posts = new Courses({coName, coTags, coIntro, coDesc, coCat, coVal, coPrice, coDisc, coImage, userId });
       const result = await posts.save();
       return NextResponse.json({result:result, success:true}, {status: 200});
 
