@@ -1,21 +1,22 @@
 import { NextResponse } from "next/server";
 import connect from "../../../../../server";
-import Profile from "../../../../../models/Profiles";
-
+import Profiles from "../../../../../models/Profiles";
 
 export const GET = async (request, {params}) => {
- try 
-  {
-
-    await connect ();
-    const posts = await  Profile.find()
-    const proById = posts.filter((item)=> item.id == params.ProId)
-    return NextResponse.json({ result:proById, success:true}, {status: 200});
-
-  } catch (error){
-    return NextResponse.json({result:"No Data Found", success:false})
+  try 
+    {
+      await connect ();
+      const profile = await Profiles.findById(params.ProId);
+      
+      if (!profile) {
+          return new NextResponse("No profile found with the given ID", {status: 404});
+      }
+      return NextResponse.json({result:profile}, {status: 200});    
+  
+    }catch(error){
+          return new NextResponse ("Erron while retrieving data: " + error, {status: 500});
+        }
   }
-}
 
 export  const  PUT = async (request, {params}) =>{ 
   try  
@@ -25,7 +26,7 @@ export  const  PUT = async (request, {params}) =>{
       const filter = {_id:profileId}
       await connect ();
 
-      const profile = await Profile.findOneAndUpdate(filter, {proName, proFather, proDob, proJob, proQual, shortIntro, proAbout, proImage, proCloc, proPloc});
+      const profile = await Profiles.findOneAndUpdate(filter, {proName, proFather, proDob, proJob, proQual, shortIntro, proAbout, proImage, proCloc, proPloc});
       return NextResponse.json({result:profile, success:true}, {status: 200});
 
     }catch(error){
