@@ -6,16 +6,16 @@ import { useRouter } from 'next/navigation';
 
 export default function UpdateUser({params}) {
 
-  const [data, setData] = useState({usrName:'', usrPass:'', usrEmail:'', usrPhone:'', usrRole:''});
+  const [data, setData] = useState({usrName:'', usrEmail:'', usrPhone:'', usrRole:''});
   const [errorMessage, setErrorMessage] = useState('');
   const router = useRouter();
   const _id = params.UserId;
 
   useEffect(() =>{
     async function fetchData() {
-      const userData = await fetch(`http://localhost:3000/api/users/${_id}`);
-      const usrById = await userData.json();
-      setData(usrById.result[0]);
+      const res = await fetch(`http://localhost:3000/api/users/${_id}`);
+      const usrById = await res.json();
+      setData(usrById.result);
     }
     fetchData();
   // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -41,10 +41,6 @@ export default function UpdateUser({params}) {
     errMsg.push('User name is required.');    
   }
   
-  if (!data.usrPass.trim()) {
-    errMsg.push('Password is required.');    
-  }
-  
   if (!data.usrEmail.trim()) {
     errMsg.push('Email is required.');    
   }
@@ -63,7 +59,7 @@ export default function UpdateUser({params}) {
     const result = await fetch (`http://localhost:3000/api/users/${_id}`, 
     {
       method:'PUT',
-      body:JSON.stringify({usrName: data.usrName, usrPass:data.usrPass, usrEmail:data.usrEmail, usrPhone:data.usrPhone, usrRole:data.usrRole}),
+      body:JSON.stringify({usrName: data.usrName,  usrEmail:data.usrEmail, usrPhone:data.usrPhone, usrRole:data.usrRole}),
     });
 
       const post = await result.json();
@@ -97,25 +93,21 @@ export default function UpdateUser({params}) {
         </div>
         <form className='p-5' onSubmit={handleSubmit}>
           <div className='flex flex-col mb-3 gap-2'>
-            <label>Change Username:</label>
-            <input type='text' name='usrName' value={data.usrName} onChange={handleChange} className='py-2 px-2 rounded-md border focus:outline-amber-600'></input>
-          </div>
-          <div className='flex flex-col mb-3 gap-2'>
-            <label>Change Password:</label>
-            <input type='password'name='usrPass' value={data.usrPass} onChange={handleChange} className='py-2 px-2 rounded-md border focus:outline-amber-600'></input>
+            <label>User Name:</label>
+            <input type='text' name='usrName' value={data?.usrName} onChange={handleChange} className='py-2 px-2 rounded-md border focus:outline-amber-600'></input>
           </div>
           <div className='flex flex-col mb-3 gap-2'>
             <label>Email Id:</label>
-            <input type='text'name='usrEmail' value={data.usrEmail} onChange={handleChange} className='py-2 px-2 rounded-md border focus:outline-amber-600'></input>
+            <input type='email'name='usrEmail' value={data?.usrEmail} onChange={handleChange} className='py-2 px-2 rounded-md border focus:outline-amber-600'></input>
           </div>
           <div className='grid grid-cols-2 gap-3'>
             <div className='flex flex-col mb-3 gap-2'>
                 <label>Phone:</label>
-                <input type='text' name='usrPhone' value={data.usrPhone} onChange={handleChange} className='py-2 px-2 rounded-md border focus:outline-amber-600'></input>
+                <input type='text' name='usrPhone' value={data?.usrPhone} onChange={handleChange} className='py-2 px-2 rounded-md border focus:outline-amber-600'></input>
             </div>
             <div className='flex flex-col mb-3 gap-2'>
                 <label>Role:</label>
-                <select disabled type='select' name='usrRole' value={data.usrRole} onChange={handleChange} className='py-2 font-bold px-2 rounded-md border focus:outline-amber-600  '>
+                <select disabled type='select' name='usrRole' value={data?.usrRole} onChange={handleChange} className='py-2 font-bold px-2 rounded-md border focus:outline-amber-600  '>
                    <option value='' className='text-center'>--- Select Role ---</option>
                    <option  value='INSTRUCTOR' className='font-bold text-sm'>INSTRUCTOR</option>
                    <option value='STUDENT' className='font-bold text-sm'>STUDENT</option>
