@@ -1,12 +1,13 @@
 'use client';
 import { toast } from "react-toastify";
 import { useRouter } from "next/navigation";
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useContext } from 'react';
 import { FaCloudUploadAlt } from "react-icons/fa";
 import Image from 'next/image';
 import { useRef } from "react";
 import React from 'react';
 import TextEditor from '@/components/TinyMce/Editor';
+import { UserContext } from "@/context/UserContext";
 
 export default function UpdateCourse({params}) {
    
@@ -16,16 +17,18 @@ export default function UpdateCourse({params}) {
     const [image, setImage] = useState('');
     const [imageData, setImageData] = useState(null); 
     const [editorContent, setEditorContent] = useState('');
+    const {loggedInUser} = useContext(UserContext);
     const [data, setData] = useState({coName:'', coTags:'', coIntro:'', coDesc:'', coPrice:'', coDisc:'', coVal:'', coCat:'', coImage:''})    
     
 
     useEffect(() =>{
       async function fetchData() {
-      let catdata = await fetch('http://localhost:3000/api/categories');
+      let catdata = await fetch('http://localhost:3000/api/categories/?userId='+ loggedInUser.result._id);
       catdata = await catdata.json();
       setCat(catdata);
     }
       fetchData();
+    // eslint-disable-next-line react-hooks/exhaustive-deps
     },[]);
 
     useEffect(() =>{
