@@ -2,11 +2,12 @@
 import TextEditor from '@/components/TinyMce/Editor';
 import { FaCloudUploadAlt } from "react-icons/fa";
 import { useRouter } from 'next/navigation';
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useContext } from 'react';
 import { toast } from 'react-toastify';
 import { useRef } from 'react';
 import Image from 'next/image';
 import React from 'react';
+import { UserContext } from '@/context/UserContext';
 
 export default function UpdateEbook({params}) {
 
@@ -15,17 +16,19 @@ export default function UpdateEbook({params}) {
     const [cat, setCat] = useState([]);
     const [image, setImage] = useState('');
     const [imageData, setImageData] = useState(null); 
+    const {loggedInUser} = useContext(UserContext);
     const [editorContent, setEditorContent] = useState(''); 
     const [errorMessage, setErrorMessage] = useState(''); 
     const [data, setData] = useState({ebkName:'', ebkTags:'', ebkIntro:'', ebkDesc:'', ebkPrice:'', ebkDisc:'', ebkAuth:'', ebkCat:'', ebkImage:''})    
 
     useEffect(() =>{
         async function fetchData() {
-          let catdata = await fetch('http://localhost:3000/api/categories');
+          let catdata = await fetch('http://localhost:3000/api/categories/?userId='+ loggedInUser.result._id);
           catdata = await catdata.json();
           setCat(catdata);
         }
         fetchData();
+      // eslint-disable-next-line react-hooks/exhaustive-deps
       },[]);
     
     useEffect(() =>{
