@@ -2,11 +2,12 @@
 import TextEditor from '@/components/TinyMce/Editor';
 import { FaCloudUploadAlt } from "react-icons/fa";
 import { useRouter } from 'next/navigation';
-import { useState, useEffect } from 'react';
+import { useState, useEffect, useContext } from 'react';
 import { toast } from 'react-toastify';
 import { useRef } from 'react';
 import Image from 'next/image';
 import React from 'react';
+import { UserContext } from '@/context/UserContext';
 
 export default function AddEbook() {
 
@@ -15,16 +16,18 @@ export default function AddEbook() {
     const [cat, setCat] = useState([]);
     const [image, setImage] = useState('');
     const [imageData, setImageData] = useState(null);  
+    const {loggedInUser} = useContext(UserContext);
     const [errorMessage, setErrorMessage] = useState(''); 
     const [data, setData] = useState({ebkName:'', ebkTags:'', ebkIntro:'', ebkDesc:'', ebkPrice:'', ebkDisc:'', ebkAuth:'', ebkCat:'', ebkImage:''})    
 
     useEffect(() =>{
         async function fetchData() {
-          let catdata = await fetch('http://localhost:3000/api/categories');
+          let catdata = await fetch('http://localhost:3000/api/categories/?userId='+ loggedInUser.result._id);
           catdata = await catdata.json();
           setCat(catdata);
         }
         fetchData();
+      // eslint-disable-next-line react-hooks/exhaustive-deps
       },[]);
 
     const showChooseFileBox = () =>{
@@ -95,7 +98,7 @@ export default function AddEbook() {
         const result = await fetch ('http://localhost:3000/api/ebooks', 
         {
           method:'POST',
-          body:JSON.stringify({ebkName:data.ebkName,  ebkTags:data.ebkTags, ebkIntro:data.ebkIntro, ebkDesc:data.ebkDesc, ebkPrice:data.ebkPrice, ebkDisc:data.ebkDisc, ebkAuth:data.ebkAuth, ebkCat:data.ebkCat, ebkImage:data.ebkImage}),
+          body:JSON.stringify({ebkName:data.ebkName,  ebkTags:data.ebkTags, ebkIntro:data.ebkIntro, ebkDesc:data.ebkDesc, ebkPrice:data.ebkPrice, ebkDisc:data.ebkDisc, ebkAuth:data.ebkAuth, ebkCat:data.ebkCat, ebkImage:data.ebkImage, userId: loggedInUser.result._id}),
         });
   
         const post = await result.json();
