@@ -1,6 +1,9 @@
-import './globals.css'
-import { Inter } from 'next/font/google'
-
+import './globals.css';
+import { Inter } from 'next/font/google';
+import { ToastContainer } from 'react-toastify';
+import 'react-toastify/dist/ReactToastify.css';
+import { GlobalState } from '@/context/UserContext';
+ 
 const inter = Inter({ subsets: ['latin'] })
 
 export const metadata = {
@@ -10,11 +13,15 @@ export const metadata = {
 
 export default function RootLayout({ children }) {
   return (
-    <html lang="en">
-      
-      <body className={inter.className}>
-        {children}
-      </body>
-    </html>
+  <html lang="en">
+    <body className={inter.className}>     
+        <ToastContainer/>
+        <GlobalState>
+          <div>
+            {children}
+          </div>
+        </GlobalState>
+     </body>
+  </html> 
   )
 }
